test NextInt

+++ src/main/java/com/mindhub/homebanking/utilitis/RandomUtil.java
@@ -9,13 +9,13 @@ public class RandomUtil {
         return "VIN-" + String.format("%0" + numDigits + "d", random.nextInt((int) Math.pow(10, numDigits)));
     }
 
-    public static long generateNumber(int numDigits) {
+    public static int generateNumber(int numDigits) {
         if (numDigits <= 0) {
             throw new IllegalArgumentException("Número de dígitos debe ser mayor que cero");
         }
 
-        long min = (long) Math.pow(10, numDigits - 1);
-        long max = (long) Math.pow(10, numDigits) - 1;
-        return min + random.nextLong(max - min + 1);
+        int min = (int) Math.pow(10, numDigits - 1);
+        int max = (int) Math.pow(10, numDigits) - 1;
+        return min + random.nextInt(max - min + 1);
     }
 }
